got mvp of the script done
