+++ src/utils/distance.ts
@@ -1,6 +1,10 @@
 import distance from "@turf/distance";
 import { point } from "@turf/helpers";
-import { Coords } from "../location-picker/LocationPicker";
+
+export interface Coords {
+  lng: number;
+  lat: number;
+}
 
 /**
  * Calculates the distance (in kilometers) between two co-ordinates.
@@ -15,30 +19,4 @@ export function getDistanceKm(coords1: Coords, coords2: Coords): number {
   const to = point([coords2.lat, coords2.lng]);
   const d = distance(from, to, { units: "kilometers" });
   return d;
-}
-
-/**
- * Formats a given distance in kilometers to meters, if the distance is
- * less than 1000 meters.
- *
- * @export
- * @param {number} km
- * @param {string} language
- * @return {string}
- */
-export function formatDistanceKm(km: number, language: string): string {
-  if (km < 1) {
-    const meters = km * 1000;
-    return new Intl.NumberFormat(language, {
-      style: "unit",
-      unit: "meter",
-      notation: "compact",
-    }).format(meters);
-  } else {
-    return new Intl.NumberFormat(language, {
-      style: "unit",
-      unit: "kilometer",
-      notation: "compact",
-    }).format(km);
-  }
 }
